feat: protect members route

+++ Backend/src/controllers/memberController.js
@@ -5,14 +5,16 @@ const {
   getMemberById,
   editMemberById,
   deleteMemberById,
+  verifyMemberAccess,
 } = require("../services/memberService");
 
 const postMemberController = async (req, res, next) => {
   try {
     MembersValidator.validateMemberPayload(req.body);
     const { name, phoneNumber } = req.body;
+    const { idOrganization } = req.credentials;
 
-    const member = await addMember({ name, phoneNumber });
+    const member = await addMember({ name, phoneNumber, idOrganization });
 
     res.status(201).json({
       status: "success",
@@ -27,7 +29,8 @@ const postMemberController = async (req, res, next) => {
 const getMembersController = async (req, res, next) => {
   try {
     const { input } = req.query;
-    const members = await getMembers(input);
+    const { idOrganization } = req.credentials;
+    const members = await getMembers(input, idOrganization);
 
     res.status(200).json({
       status: "success",
@@ -43,8 +46,9 @@ const getMembersController = async (req, res, next) => {
 const getMemberByIdController = async (req, res, next) => {
   try {
     const { id } = req.params;
+    const { idOrganization } = req.credentials;
 
-    const member = await getMemberById(id);
+    const member = await getMemberById(id, idOrganization);
 
     res.status(200).json({
       status: "success",
@@ -61,9 +65,10 @@ const putMemberByIdController = async (req, res, next) => {
   try {
     MembersValidator.validateMemberPayload(req.body);
     const { id } = req.params;
-    const { name, phoneNumber } = req.body;
+    const { idOrganization } = req.credentials;
 
-    await editMemberById(id, { name, phoneNumber });
+    await verifyMemberAccess(id, idOrganization);
+    await editMemberById(id, req.body);
 
     res.status(200).json({
       status: "success",
@@ -77,7 +82,9 @@ const putMemberByIdController = async (req, res, next) => {
 const deleteMemberByIdController = async (req, res, next) => {
   try {
     const { id } = req.params;
+    const { idOrganization } = req.credentials;
 
+    await verifyMemberAccess(id, idOrganization);
     await deleteMemberById(id);
 
     res.status(200).json({
